minor cleanings to allow inclusion in c++ projects
also minor changes to shut up pedantic static analyzers :
- no more double underscores as prefix in identifiers
- checking the return value from calls to the pthread interface

Note that there isn't much that can be done if calls to the pthread
interface fails (it can happen only in case of memory coruuption or if
there is a disastrous bug in the logic). The choice here was to call
abort() and let the caller the choice to catch the signal and take more actions
or to let the process die with a crash dump.